Rename schemas/puzzles -> puzzle
