+++ imports/client/components/Puzzle.tsx
@@ -10,7 +10,7 @@ import React, {
 import Button from 'react-bootstrap/Button';
 import { Link } from 'react-router-dom';
 import Ansible from '../../ansible';
-import { PuzzleType } from '../../lib/schemas/puzzles';
+import { PuzzleType } from '../../lib/schemas/puzzle';
 import { TagType } from '../../lib/schemas/tags';
 import PuzzleAnswer from './PuzzleAnswer';
 import PuzzleModalForm, { PuzzleModalFormSubmitPayload } from './PuzzleModalForm';
