+++ imports/client/components/PuzzleListPage.tsx
@@ -26,7 +26,7 @@ import Puzzles from '../../lib/models/puzzles';
 import Tags from '../../lib/models/tags';
 import { userMayWritePuzzlesForHunt } from '../../lib/permission_stubs';
 import { HuntType } from '../../lib/schemas/hunt';
-import { PuzzleType } from '../../lib/schemas/puzzles';
+import { PuzzleType } from '../../lib/schemas/puzzle';
 import { TagType } from '../../lib/schemas/tags';
 import PuzzleList from './PuzzleList';
 import PuzzleModalForm, {
