+++ imports/client/components/RTCDebugPage.tsx
@@ -14,7 +14,7 @@ import Puzzles from '../../lib/models/puzzles';
 import { CallParticipantType } from '../../lib/schemas/call_participant';
 import { CallSignalType } from '../../lib/schemas/call_signal';
 import { ProfileType } from '../../lib/schemas/profile';
-import { PuzzleType } from '../../lib/schemas/puzzles';
+import { PuzzleType } from '../../lib/schemas/puzzle';
 
 interface ParticipantSummaryProps {
   displayName: string;
