+++ imports/client/components/RelatedPuzzleList.tsx
@@ -1,6 +1,6 @@
 import { _ } from 'meteor/underscore';
 import React from 'react';
-import { PuzzleType } from '../../lib/schemas/puzzles';
+import { PuzzleType } from '../../lib/schemas/puzzle';
 import { TagType } from '../../lib/schemas/tags';
 import PuzzleList from './PuzzleList';
 import { puzzleInterestingness } from './puzzle-sort-and-group';
