+++ imports/lib/schemas/facade.ts
@@ -11,7 +11,7 @@ import Hunt from './hunt';
 import PendingAnnouncement from './pending_announcement';
 import Profile from './profile';
 import PublicSetting from './public_setting';
-import Puzzles from './puzzles';
+import Puzzle from './puzzle';
 import Tags from './tags';
 import User from './users';
 
@@ -29,7 +29,7 @@ const Schemas = {
   PendingAnnouncement,
   Profile,
   PublicSetting,
-  Puzzles,
+  Puzzle,
   Tags,
   User,
 };
