+++ imports/model-helpers.ts
@@ -4,7 +4,7 @@ import { _ } from 'meteor/underscore';
 import Mustache from 'mustache';
 import MeteorUsers from './lib/models/meteor_users';
 import { HuntType } from './lib/schemas/hunt';
-import { PuzzleType } from './lib/schemas/puzzles';
+import { PuzzleType } from './lib/schemas/puzzle';
 
 const answerify = function (answer: string): string {
   return answer.toUpperCase();
