+++ imports/server/chat-notifications.ts
@@ -7,7 +7,7 @@ import Hunts from '../lib/models/hunts';
 import Puzzles from '../lib/models/puzzles';
 import { ChatNotificationType } from '../lib/schemas/chat_notification';
 import { HuntType } from '../lib/schemas/hunt';
-import { PuzzleType } from '../lib/schemas/puzzles';
+import { PuzzleType } from '../lib/schemas/puzzle';
 import RefCountedObserverMap from './refcounted-observer-map';
 
 class ChatNotificationWatcher {
